git-svn-id: https://db-svn.informatik.uni-bremen.de/repos/stuff/use_plugins@5672 e95982d2-0d12-e879-a455-8ce5a0885042

+++ ModelValidator/branches/diplom-aili/model-validator/src/org/tzi/kodkod/model/config/impl/DefaultConfigurationValues.java
@@ -26,9 +26,9 @@ public class DefaultConfigurationValues {
 	public static final int integerMin = -10;
 	public static final int integerMax = 10;
 
-	public static final int realMin = -2;
-	public static final int realMax = 2;
-	public static final double realStep = 0.5;
+	public static final double realMin = -2d;
+	public static final double realMax = 2d;
+	public static final double realStep = 0.5d;
 	
 	public static final boolean AGGREGATIONCYCLEFREENESS = false;
 	public static final boolean FORBIDDENSHARING = true;
